priorityをuint32_tに変更

+++ project/Engine/Component/Components/SpriteRenderer/SpriteRendererComponent.h
@@ -117,7 +117,7 @@ public:
 
 	void SetAnchor(const Vector2f& anchor) { anchor_ = anchor; }
 	void SetPivot(const Vector2f& pivot) { pivot_ = pivot; }
-	void SetPriority(float priority) { priority_ = priority; }
+	void SetPriority(uint32_t priority) { priority_ = priority; }
 
 	//* render option *//
 
@@ -169,7 +169,7 @@ private:
 	std::bitset<2> isFlip2d_ = {};
 	std::bitset<2> isFlipUV_ = {};
 
-	float priority_ = 0.0f;
+	uint32_t priority_ = 0;
 
 	//* triangle input assembler *//
